workers/requirements.txt
requests
python-dotenv
sqlalchemy
pymysql
